SD card emulation now can perform reads.
Since there is nothing to back it, currently it will simply return zeros
everywhere. This makes the kernel's sdBase.S choke on it as it expects a
nonzero value when hitting the sd_find_div_r2425_by_r23_result_r0 routine
while trying to interpret the MBR and the FAT. But it reaches there
correctly (issuing a CMD18, stopping it with a CMD12, then another CMD18).