Bugs fixed in SD interface related to timing.
Cycle event goals were set incorrectly resulting in the SD card unable to init
(in reasonable time) or dropping phases of operation. Now that these are fixed
Toorum's Quest properly goes through a CMD24 write (which due to the stub in
the virtual FAT is not actually written, but at least it works).