Instruction deflagger added, providing some speed bump.
Somewhat experimental, but in theory it should function proper. It does a pass
on the compiled instruction stream proving whether the flags of arithmetic
instructions were used or not, and replaces them to variants which don't
calculate flags if they don't.

For native compiles the result might be around 0 - 10 percents, a bit dubious
(the greatest bottleneck is likely the 28 millions branch mispredictions per
second for interpreting the instruction stream). Emscripten receives the
greatest benefit since it is probably impossible to optimize well enough the
complex flag logic through the many passes until the program gets to start
running in a browser.