Added flag precalculation to boost AVR emulation.
This is partially a trade-off between instruction cache & processing and data
cache as to store the precalculated flags a little more than 128K of data is
required.

By experience it does not result in anything running slower than before while
many games run about 10-20% faster. The performance boost is similar in
Emscripten.