Preparations for emulator state saves.
The code ROM and the Virtual FAT filesystem got an access indicator to
signal whether they have to be included in a state save. Other components
doesn't necessarily need this (the SPI RAM's contents might simply be RLE
compressed to prevent it adding 128K of void if it is not used).