Added frame merging capability.
The frame merging merges the current display frame with the previous one,
which can smoothen the display of certain sprite cycling algorithms
(translucent sprites instead of flickering). It makes the emulator running a
little slower.

The base performance of the renderer was improved meanwhile, making the
emulator running a few percents faster with frame merging off.