Restored ability to output to title & console along with optimizations.
The optimizations only cover reducing the amount of graphics updates and some
info preparations for those when it is known that an area is not visible. The
effect is negligible (compared to the other loads).

Also added a flag to remove all console output which in some cases seem to
severely effect performance (the console output slowing the emulator down
slightly).